Moved Application Choices controller
